+++ Entity/Paciente.cs
@@ -6,18 +6,18 @@ namespace Entity
     {
         public string Identificacion{ get; set; } 
         public string Nombre{ get; set; }
-        public double ValorHospital{ get; set; }
-        public double SalarioTrabajador{ get; set; }
-        public double ValorCopago{ get{
-            double total,Porcentaje = 0;
+        public decimal ValorHospital{ get; set; }
+        public decimal SalarioTrabajador{ get; set; }
+        public decimal ValorCopago{ get{
+            decimal total,Porcentaje = 0;
          if(SalarioTrabajador>=25000000)
          {
-             Porcentaje=0.20;
+             Porcentaje=0.20m;
             total=(this.ValorHospital*Porcentaje);
          }
          else
          {
-              Porcentaje=0.10;
+              Porcentaje=0.10m;
             total=(this.ValorHospital*Porcentaje);  
          }
          return total;
